requirements.txt
torch==1.6.0
numpy==1.19.1
scipy==1.5.2
matplotlib
Pillow
